Tightened code in pc_fallback.py
Unnecessary terminal output removed and check against temporary table
returns object direct rather than referencing the individual attributes.
