+++ .bash_history
@@ -169,3 +169,122 @@ exit
 export POSTCODEINFO_AUTH_TOKEN=a4a3c5ed091552daf8d4a6b87a43bf100e359e89
 ./runserver.sh
 exit
+./runonce.sh 
+./runserver.sh 
+exit
+export POSTCODEINFO_AUTH_TOKEN=a4a3c5ed091552daf8d4a6b87a43bf100e359e89
+./runserver.sh 
+exit
+./runserver.sh 
+exit
+export POSTCODEINFO_AUTH_TOKEN=a4a3c5ed091552daf8d4a6b87a43bf100e359e89
+./runserver.sh 
+exit
+export POSTCODEINFO_AUTH_TOKEN=a4a3c5ed091552daf8d4a6b87a43bf100e359e89
+./runserver.sh 
+cd laalaa/apps/advisers/tests/
+ls -la
+python test_geocoder.py 
+export DJANGO_SETTINGS_MODULE=laalaa.settings
+python test_geocoder.py 
+export DJANGO_SETTINGS_MODULE=/home/app/laalaa/settings/base.py
+python test_geocoder.py 
+export DJANGO_SETTINGS_MODULE=laalaa.settings.base
+python test_geocoder.py 
+export DJANGO_SETTINGS_MODULE=laalaa.settings
+python test_geocoder.py 
+cd ..
+cd ..
+ls -la
+cd ..
+ls -la
+cd ..
+django-admin runserver --settings=laalaa.settings
+cd /home/app/
+ls -la
+pwd
+pwd
+cd laalaa/apps/advisers/
+ls -la
+python pc_test.py 
+python pc_test.py 
+python -m unittest pc_test.py 
+exit
+python manage.py test
+python manage.py test laalaa/apps/advisers
+python manage.py test laalaa.apps.advisers.tests.test_pc_fallback
+python manage.py test laalaa.apps.advisers.tests
+exit
+python manage.py test laalaa.apps.advisers.tests
+python manage.py test laalaa.apps.advisers.tests
+exit
+python manage.py test laalaa.apps.advisers.tests.test_pc_fallback
+exit
+python manage.py test laalaa.apps.advisers.tests.test_pc_fallback
+python manage.py test laalaa.apps.advisers.tests
+exit
+python manage.py test laalaa.apps.advisers.tests.test_geocode_fallback
+python manage.py test laalaa.apps.advisers.tests.test_pc_fallback.test_geocode_Fallback
+python manage.py test laalaa.apps.advisers.tests.test_pc_fallback.test_geocode_fallback
+python manage.py test advisers.tests
+python manage.py test advisers.tests
+python manage.py test advisers.tests.test_pc_fallback
+exit
+export POSTCODEINFO_AUTH_TOKEN=a4a3c5ed091552daf8d4a6b87a43bf100e359e89
+python manage.py test advisers.tests.test_pc_fallback
+python manage.py test advisers.tests.test_pc_fallback
+python manage.py test advisers.tests.test_geocode_fallback
+python manage.py test advisers.tests.test_pc_fallback.test_geocode_fallback
+python
+python
+pwd
+python
+python manage.py shell
+pwd
+python manage.py test laalaa/apps/advisers/fbtest/
+python manage.py test advisers.tests.test_pc_fallback.test_geocode_fallback
+python manage.py test advisers.tests.test_pc_fallback.test_geocode_fallback
+exit
+python manage.py test advisers.tests.test_pc_fallback.test_geocode_fallback
+exit
+./runserver.sh 
+python manage.py tets
+python manage.py test
+exit
+ls -la
+python manage.py test laalaa.apps.advisers.tests.test_pc_fallback:FallbackTest.test_geocode_fallback
+exit
+./runserver.sh 
+exit
+./runserver.sh 
+./runonce.sh 
+./runserver.sh 
+export POSTCODEINFO_AUTH_TOKEN=a4a3c5ed091552daf8d4a6b87a43bf100e359e89
+./runserver.sh 
+ecit
+exit
+export POSTCODEINFO_AUTH_TOKEN=a4a3c5ed091552daf8d4a6b87a43bf100e359e89
+./runserver.sh 
+exit
+export POSTCODEINFO_AUTH_TOKEN=a4a3c5ed091552daf8d4a6b87a43bf100e359e89
+./runserver.sh 
+exit
+export POSTCODEINFO_AUTH_TOKEN=a4a3c5ed091552daf8d4a6b87a43bf100e359e89
+./runserver.sh 
+exit
+./runserver.sh 
+exit
+export POSTCODEINFO_AUTH_TOKEN=a4a3c5ed091552daf8d4a6b87a43bf100e359e89
+./runserver.sh 
+exit
+./runserver.sh 
+export POSTCODEINFO_AUTH_TOKEN=a4a3c5ed091552daf8d4a6b87a43bf100e359e89
+./runonce.sh 
+./runserver.sh 
+exit
+export POSTCODEINFO_AUTH_TOKEN=a4a3c5ed091552daf8d4a6b87a43bf100e359e89
+./runserver.sh 
+exit
+export POSTCODEINFO_AUTH_TOKEN=a4a3c5ed091552daf8d4a6b87a43bf100e359e89
+./runserver.sh 
+exit
