+++ requirements/base.txt
@@ -14,4 +14,6 @@ postcodeinfo>=0.0.2,<0.1.0
 postcodes
 django-extensions
 werkzeug
+mock
+django-nose
